+++ src/app/components/Input/Input.actions.js
@@ -1,0 +1,6 @@
+export const addTodo = (todo) => {
+    return {
+        type: 'ADD_TODO',
+        todo
+    }
+};
