+++ src/app/components/Input/Input.actions.test.js
@@ -1,20 +1,9 @@
-import { addTodo, toggleTodoState } from './actions';
+import { addTodo } from './Input.actions';
 
-describe('Actions: App', () => {
+describe('Actions: InputComponent', () => {
     test('should be imported', () => {
         expect(addTodo).toBeDefined();
         expect(typeof addTodo).toBe('function');
-        expect(typeof toggleTodoState).toBe('function');
-        expect(toggleTodoState).toBeDefined();
-    });
-
-    describe('toggleTodoState () =>', () => {
-        test('should toggle todo item status', () => {
-            expect(toggleTodoState(2)).toEqual({
-                type: 'TODO_TOGGLE_DONE',
-                id: 2
-            });
-        });
     });
 
     describe('addTodo () =>', () => {
