+++ src/app/styles.js
@@ -141,42 +141,5 @@ export default StyleSheet.create({
         width: '20px',
         height: '20px',
         color: 'black'
-    },
-    iconAddTodoButton: {
-        color: 'rgb(193, 193, 193)',
-        margin: '10px 0'
-    },
-    fieldSelectSeverity: {
-        position: 'relative',
-        width: '120px',
-        minWidth: '120px',
-        height: '30px',
-        right: '130px',
-        bottom: '-12.5px',
-        overflow: 'hidden'
-    },
-    textSeveritySelected: {
-        textAlign: 'left',
-        padding: '0 10px 0',
-        margin: '0',
-        fontSize: '1rem',
-        color: 'rgba(102, 102, 102, .6)'
-    },
-    selectSeverity: {
-        width: '100%',
-        height: '100%',
-        position: 'absolute',
-        left: '0',
-        top: '0',
-        margin: '0',
-        padding: '0',
-        borderRadius: '4px',
-        background: 'transparent',
-        color: 'transparent',
-        border: 'none',
-        outline: 'none',
-        display: 'inline-block',
-        appearance: 'none',
-        cursor: 'pointer'
     }
 });
